requirements.txt
streamlit
google-generativeai
python-dotenv
requests
beautifulsoup4
langchain-text-splitters